% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_auth.R
\name{monday_auth}
\alias{monday_auth}
\title{Authorize Monday API with your token}
\usage{
monday_auth(
  client_id = Sys.getenv("MONDAY_CLIENT_ID"),
  client_secret = Sys.getenv("MONDAY_CLIENT_SECRET"),
  token = NULL,
  new_user = FALSE,
  silent = FALSE
)
}
\description{
Enter your token that can be found on https://app.monday.com/app/profile.
}
\examples{
monday_auth(client_id = Sys.getenv("MONDAY_CLIENT_ID"), client_secret = Sys.getenv("MONDAY_CLIENT_SECRET"), token = NULL, new_user = FALSE, silent = FALSE)
}
