% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_column_values.R
\name{monday_column_values}
\alias{monday_column_values}
\title{Get column values}
\usage{
monday_column_values(
  item_id = NULL,
  board_id = NULL,
  column_id = NULL,
  column_title = NULL,
  column_type = NULL,
  raw = FALSE
)
}
\arguments{
\item{item_id}{The id of a Monday.com item (optional)}

\item{board_id}{The id of a Monday.com board (optional)}

\item{column_id}{The id of a Monday.com board column (optional)}

\item{column_title}{The title of a Monday.com board column (optional)}

\item{column_type}{The type of a Monday.com board column (optional)}

\item{raw}{Wether or not to return raw or structured data (defaults to FALSE)}
}
\description{
Get column values for Monday.com items
}
\examples{
monday_column_values(item_id = 123, board_id = 1234, column_title="Status")
}
