% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_list_boards.R
\name{monday_list_boards}
\alias{monday_list_boards}
\title{A data frame of Monday.com boards}
\usage{
monday_list_boards()
}
\description{
Get a list of all boards the aunthenticated user has access to
}
\examples{
monday_list_boards()
}
