% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_list_columns.R
\name{monday_list_columns}
\alias{monday_list_columns}
\title{A data frame of columns}
\usage{
monday_list_columns(board_id)
}
\arguments{
\item{board_id}{The id of a Monday.com board}
}
\description{
Get the columns of a Monday.com board
}
\examples{
monday_list_columns(board_id = 1234)
}
