% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_list_items.R
\name{monday_list_items}
\alias{monday_list_items}
\title{Get items from Monday.com}
\usage{
monday_list_items(board_id = NULL, limit = NULL, custom_item_query = NULL)
}
\arguments{
\item{board_id}{The id of a Monday.com board (optional)}

\item{limit}{Choose a maximum number of items to get (optional)}

\item{custom_item_query}{Provide your own item query (optional)}
}
\description{
A data frame of Monday.com items/pulses
}
\examples{
monday_list_items(board_id = 1234, limit = 10, custom_item_query="{items{name}}")
}
