% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_time_tracking.R
\name{monday_time_tracking}
\alias{monday_time_tracking}
\title{Get time tracking data}
\usage{
monday_time_tracking(board_id = NULL, raw = FALSE)
}
\arguments{
\item{board_id}{The id of a Monday.com board (optional)}

\item{raw}{Wether or not to return raw or structured data (defaults to FALSE)}
}
\description{
Get time tracking data from your Monday.com boards
}
\examples{
monday_time_tracking(board_id = 1234)
}
