% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_query.R
\name{monday_query}
\alias{monday_query}
\title{Get custom API request string}
\usage{
monday_query(query)
}
\arguments{
\item{query}{Query parameters for the Graph API}
}
\description{
Type in your own request string
}
\examples{
monday_query(query)
}
